Model Cache Fix
